LPS-162251 Don't over componentize until there's value
